+++ internal/git/log/log.go
@@ -3,13 +3,13 @@ package log
 import (
 	"bufio"
 	"bytes"
+	"context"
 	"fmt"
 	"io"
 	"strings"
 
-	"gitlab.com/gitlab-org/gitaly/internal/git"
-
 	pb "gitlab.com/gitlab-org/gitaly-proto/go"
+	"gitlab.com/gitlab-org/gitaly/internal/git"
 )
 
 // Parser holds necessary state for parsing a git log stream
@@ -18,16 +18,24 @@ type Parser struct {
 	currentCommit *pb.GitCommit
 	finished      bool
 	err           error
+	*commitMessageHelper
 }
 
 const fieldDelimiter = "\x1f"
 
 // NewLogParser returns a new Parser
-func NewLogParser(src io.Reader) *Parser {
-	parser := &Parser{}
-	parser.reader = bufio.NewReader(src)
+func NewLogParser(ctx context.Context, repo *pb.Repository, src io.Reader) (*Parser, error) {
+	cmh, err := newCommitMessageHelper(ctx, repo)
+	if err != nil {
+		return nil, err
+	}
 
-	return parser
+	parser := &Parser{
+		reader:              bufio.NewReader(src),
+		commitMessageHelper: cmh,
+	}
+
+	return parser, nil
 }
 
 // Parse parses a single git log line. It returns true if successful, false if it finished
@@ -54,19 +62,25 @@ func (parser *Parser) Parse() bool {
 	}
 
 	elements := bytes.Split(line, []byte(fieldDelimiter))
-	if len(elements) != 10 {
+	if len(elements) != len(commitLogFormatFields) {
 		parser.err = fmt.Errorf("error parsing ref: %q", line)
 		return false
 	}
 
-	var parentIds []string
-	if len(elements[9]) > 0 {
-		parentIds = strings.Split(string(elements[9]), " ")
+	subject, body, err := parser.commitMessage(string(elements[0]))
+	if err != nil {
+		parser.err = err
+		return false
 	}
 
-	commit, err := git.NewCommit(elements[0], elements[1], elements[2],
-		elements[3], elements[4], elements[5], elements[6], elements[7],
-		elements[8], parentIds...)
+	var parentIds []string
+	if parentFieldIndex := len(commitLogFormatFields) - 1; len(elements[parentFieldIndex]) > 0 {
+		parentIds = strings.Split(string(elements[parentFieldIndex]), " ")
+	}
+
+	commit, err := git.NewCommit(elements[0], []byte(subject), []byte(body),
+		elements[1], elements[2], elements[3], elements[4], elements[5],
+		elements[6], parentIds...)
 	if err != nil {
 		parser.err = err
 		return false